refactor(examples): Reorganize OTel tests
